[EA3-123] refactor:예외처리가 ValidException을 상속받도록 변경
