+++ src/main/java/grep/neogul_coder/domain/recruitment/RecruitmentErrorCode.java
@@ -1,6 +1,6 @@
 package grep.neogul_coder.domain.recruitment;
 
-import grep.neogul_coder.global.response.ErrorCode;
+import grep.neogul_coder.global.code.ErrorCode;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
