+++ src/main/java/grep/neogul_coder/global/auth/jwt/JwtExceptionFilter.java
@@ -3,8 +3,8 @@ package grep.neogul_coder.global.auth.jwt;
 import grep.neogul_coder.global.exception.AuthApiException;
 import grep.neogul_coder.global.exception.AuthWebException;
 import grep.neogul_coder.global.exception.CommonException;
-import grep.neogul_coder.global.response.Code;
-import grep.neogul_coder.global.response.CommonCode;
+import grep.neogul_coder.global.code.Code;
+import grep.neogul_coder.global.code.CommonCode;
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
