+++ src/main/java/grep/neogul_coder/global/code/Code.java
@@ -1,4 +1,4 @@
-package grep.neogul_coder.global.response;
+package grep.neogul_coder.global.code;
 
 import org.springframework.http.HttpStatus;
