+++ src/main/java/grep/neogul_coder/global/code/ErrorCode.java
@@ -1,4 +1,4 @@
-package grep.neogul_coder.global.response;
+package grep.neogul_coder.global.code;
 
 public interface ErrorCode extends Code {
 }
