+++ src/main/java/grep/neogul_coder/global/exception/AuthApiException.java
@@ -1,6 +1,6 @@
 package grep.neogul_coder.global.exception;
 
-import grep.neogul_coder.global.response.Code;
+import grep.neogul_coder.global.code.Code;
 
 public class AuthApiException extends CommonException {
