+++ src/main/java/grep/neogul_coder/global/exception/advice/AuthExceptionAdvice.java
@@ -3,7 +3,7 @@ package grep.neogul_coder.global.exception.advice;
 import grep.neogul_coder.global.exception.AuthApiException;
 import grep.neogul_coder.global.exception.GoogleUserLoginException;
 import grep.neogul_coder.global.response.ApiResponse;
-import grep.neogul_coder.global.response.CommonCode;
+import grep.neogul_coder.global.code.CommonCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
