+++ src/main/java/grep/neogul_coder/global/exception/business/BusinessException.java
@@ -1,6 +1,6 @@
 package grep.neogul_coder.global.exception.business;
 
-import grep.neogul_coder.global.response.ErrorCode;
+import grep.neogul_coder.global.code.ErrorCode;
 
 public class BusinessException extends RuntimeException {
