+++ src/main/java/grep/neogul_coder/global/exception/business/NotFoundException.java
@@ -1,6 +1,6 @@
 package grep.neogul_coder.global.exception.business;
 
-import grep.neogul_coder.global.response.ErrorCode;
+import grep.neogul_coder.global.code.ErrorCode;
 
 public class NotFoundException extends BusinessException{
     public NotFoundException(ErrorCode errorcode, String message) {
