+++ src/main/java/grep/neogul_coder/global/exception/validation/DuplicatedException.java
@@ -1,6 +1,6 @@
 package grep.neogul_coder.global.exception.validation;
 
-import grep.neogul_coder.global.response.ErrorCode;
+import grep.neogul_coder.global.code.ErrorCode;
 
 public class DuplicatedException extends ValidationException {
     public DuplicatedException(ErrorCode errorCode) {
